+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Fixed
+- Readonly select questions still selectable.
+
 [4.7.8] - 2016-05-17
 ---------------------
 ##### Fixed